# src/main/java/io/swagger/api/ProvinciasApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import java.util.List;

import javax.annotation.Generated;
import javax.validation.Valid;
import javax.validation.constraints.Max;
import javax.validation.constraints.Min;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import io.swagger.model.Poblacion;
import io.swagger.model.Provincia;

@Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-06-12T19:52:47.566Z")

@Api(value = "provincias", description = "Consultar provincias y poblaciones", tags="Provincias")
public interface ProvinciasApi {

    @ApiOperation(value = "Consultar poblaciones", nickname = "getPoblaciones", notes = "Devuelve un listado de poblaciones que comiencen por la cadena de busqueda, con paginacion. ", response = Poblacion.class, responseContainer = "List", tags={ "Provincias", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Poblacion.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Peticion incorrecta", response = Error.class),
        @ApiResponse(code = 404, message = "No encontrado", response = Error.class) })
    @RequestMapping(value = "/provincias/{idProvincia}/poblaciones",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Poblacion>> getPoblaciones(@ApiParam(value = "ID de la provincia para recuperar sus poblaciones",required=true) @PathVariable("idProvincia") Integer idProvincia,@ApiParam(value = "Pasa una cadena opcional para buscar poblaciones") @Valid @RequestParam(value = "searchString", required = false) String searchString,@Min(0)@ApiParam(value = "Numero de pagina solicitada (desde 0)", defaultValue = "0") @Valid @RequestParam(value = "page", required = false, defaultValue="0") Integer page,@Min(0) @Max(50) @ApiParam(value = "Numero registros por pagina", defaultValue = "5") @Valid @RequestParam(value = "size", required = false, defaultValue="5") Integer size);


    @ApiOperation(value = "Consulta provincias", nickname = "getProvincias", notes = "Devuelve un listado de provincias en funcion de un parametro de busqueda opcional. ", response = Provincia.class, responseContainer = "List", tags={ "Provincias", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Provincia.class, responseContainer = "List"),
        @ApiResponse(code = 404, message = "No encontrado", response = Error.class) })
    @RequestMapping(value = "/provincias",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Provincia>> getProvincias(@ApiParam(value = "Pasa una cadena opcional para buscar provincias", defaultValue = "") @Valid @RequestParam(value = "searchString", required = false, defaultValue="") String searchString);

}